pyproject.toml
[project]
name = "local-fpt"
version = "0.1.0"
description = "Local Flow Production Tracking (ShotGrid) backend for development"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Local FPT Team", email = "dev@localfpt.local" }
]
dependencies = [
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
validation = [
    "shotgun-api3>=0.13.0",
]

[project.scripts]
local-fpt = "local_fpt.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/local_fpt"]
branch = true
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
src = ["src"]
target-version = "py310"
line-length = 100
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"tests/*" = ["B011"]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false

[tool.hatch.build.targets.wheel]
packages = ["src/local_fpt"]
